requirements.txt
pytorch
matplotlib
numpy
pandas
torchtext
dill
spacy